requirements.txt
flask
psycopg[binary]
gunicorn
psycogy-pool
python-dotenv


